+++ BeeDance.h
@@ -99,7 +99,7 @@ class BeeDanceTracker : public BioTracker::Core::TrackingAlgorithm {
     void switchMode(bool atracking);
     bool clickInsideRectangle(std::vector<cv::Point2i> pts, QMouseEvent *e);
     void drawPath(QPainter *painter);
-    void drawRectangle(QPainter *painter, int frame);
-    void updatePoints(int frame);
-    std::vector<QPointF> getArrowPoints(int frame, int cto);
+    void drawRectangle(QPainter *painter, size_t frame);
+    void updatePoints(size_t frame);
+    std::vector<QPointF> getArrowPoints(size_t frame, size_t cto);
 };
